requirements.txt
python-telegram-bot==4.2.1
mysql-connector
